order + payment

+++ src/main/java/com/example/tacoshop/dto/request/ProductRequest.java
@@ -5,13 +5,15 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public record ProductRequest(
         @NotBlank(message = "Product name is required")
         @Size(min = 2, max = 100, message = "Product name must be between 2 and 100 characters")
         String name,
         @NotNull(message = "Price is required")
         @Positive(message = "Price must be positive")
-        Long price,
+        BigDecimal price,
         @NotNull(message = "Stock is required")
         @Positive(message = "Stock must be positive")
         Integer stock,
