+++ src/main/java/com/example/tacoshop/dto/response/CustomerProductResponse.java
@@ -1,9 +1,11 @@
 package com.example.tacoshop.dto.response;
 
+import java.math.BigDecimal;
+
 public record CustomerProductResponse(
         Long id,
         String name,
-        Long price,
+        BigDecimal price,
         Integer stock
 ) {
 }
