+++ src/main/java/com/example/tacoshop/dto/response/OrderItemResponse.java
@@ -1,10 +1,12 @@
 package com.example.tacoshop.dto.response;
 
+import java.math.BigDecimal;
+
 public record OrderItemResponse(
         Long productId,
         String productName,
         Integer quantity,
-        Long priceAtOrder,
+        BigDecimal priceAtOrder,
         String customData
 ) {
 }
