+++ src/main/java/com/example/tacoshop/dto/response/OrderResponse.java
@@ -1,13 +1,14 @@
 package com.example.tacoshop.dto.response;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.List;
 
 public record OrderResponse(
         UserResponse customer,
         List<OrderItemResponse> orderItems,
-        Long totalAmount,
-        Long discountAmount,
+        BigDecimal totalAmount,
+        BigDecimal discountAmount,
         String discountCode,
         String status,
         OffsetDateTime expiresAt
