+++ src/main/java/com/example/tacoshop/entity/OrderItem.java
@@ -7,6 +7,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "order_items")
 @SuperBuilder
@@ -19,13 +21,17 @@ public class OrderItem extends BaseModel<Long> {
     @ManyToOne
     @JoinColumn(name = "order_id", nullable = false)
     private OrderEntity order;
+
     @ManyToOne
     @JoinColumn(name = "product_id", nullable = false)
     private Product product;
+
     @Column(nullable = false)
     private Integer quantity;
+
     @Column(nullable = false)
-    private Long priceAtOrder;
+    private BigDecimal priceAtOrder;
+
     @Lob
     @Column(name = "large_text", columnDefinition = "CLOB")
     private String customData;
